chore: rename ExploreTabs
